cli: cli to separate file
